[add] sumit added Blazor with all CRUD operation

+++ CommonLayer/Model/EmployeeModel.cs
@@ -10,7 +10,7 @@ namespace CommonLayer.Model
     {
         [Key]
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public long EmployeeId { get; set; }
+        public int EmployeeId { get; set; }
         [Required]
         public string Name { get; set; }
         [Required]
